requirements/main.txt
discord.py >= 2.6.0
aiohttp >= 3.9.3
Pillow >= 8.0
aiosqlite >= 0.17.0
docx2python >= 1.27.0
python-dateutil >= 2.8.0
jishaku >= 2.6.0
beautifulsoup4 >= 4.12.2
thefuzz >= 0.22.1
